Refactor reservationPopup and commentPopup
