+++ src/index.html
@@ -31,7 +31,6 @@
 
     <main>
       <div id="popWindow"></div>
-      <div id="reservationWindow"></div>
       <div class="movie-container"></div>
     </main>
