+++ src/index.js
@@ -1,14 +1,14 @@
 import "./styles/style.css";
 import showMovies from "./modules/api.js";
-import { showPopup, comments } from "./modules/commentPopup.js";
-import { createReservationPopup, reservations } from "./modules/reservationPopup.js";
+import popupWin from "./modules/pop";
+import { comments } from "./modules/commentPopup.js";
+import { reservations } from "./modules/reservationPopup.js";
 
 window.addEventListener("load", () => {
   showMovies();
-  const reservationWindow = document.getElementById("reservationWindow");
-  reservationWindow.appendChild(createReservationPopup());
-  
+
+  const popData = document.getElementById("popWindow");
+  popData.appendChild(popupWin());
   document.addEventListener("click", reservations);
-  document.getElementById('popWindow').innerHTML = showPopup();
   document.addEventListener("click", comments);
 });
