+++ src/styles/style.css
@@ -166,7 +166,7 @@ footer {
 }
 
 .titleContent h2 {
-  font-family: 'Crete Round', serif;
+  font-family: "Crete Round", serif;
   font-style: normal;
   font-weight: 400;
   font-size: 32px;
